fix(#6217): unpack jars through an Unpacked object
Static unpack on CentralMaven was the wrong shape for this codebase.
Move the Zip Slip-safe extraction into Unpacked, a small Scalar that
owns the jar and destination, and have CentralMaven compose it.
