+++ eo-maven-plugin/src/main/java/org/eolang/maven/CentralMaven.java
@@ -6,16 +6,12 @@ package org.eolang.maven;
 
 import com.jcabi.log.Logger;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardCopyOption;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.BiConsumer;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
 import org.apache.maven.model.Dependency;
 import org.apache.maven.repository.internal.MavenRepositorySystemUtils;
 import org.eclipse.aether.DefaultRepositorySystemSession;
@@ -182,7 +178,7 @@ final class CentralMaven implements BiConsumer<Dependency, Path> {
             artifact.getFile()
         );
         try {
-            CentralMaven.unpack(artifact.getFile().toPath(), dest);
+            new Unpacked(artifact.getFile().toPath(), dest).value();
         } catch (final IOException ex) {
             throw new IllegalStateException(
                 String.format(
@@ -210,40 +206,6 @@ final class CentralMaven implements BiConsumer<Dependency, Path> {
         final BiConsumer<? super Dependency, ? super Path> after
     ) {
         throw new UnsupportedOperationException("not implemented #andThen()");
-    }
-
-    /**
-     * Unpacks a JAR (ZIP) file into the given directory.
-     * Rejects entries whose resolved path would leave {@code dest} (Zip Slip).
-     * @param jar Path to the JAR file
-     * @param dest Destination directory
-     * @throws IOException If unpacking fails or an entry escapes {@code dest}
-     */
-    static void unpack(final Path jar, final Path dest) throws IOException {
-        final Path root = dest.toAbsolutePath().normalize();
-        Files.createDirectories(root);
-        try (ZipInputStream zis = new ZipInputStream(Files.newInputStream(jar))) {
-            ZipEntry entry = zis.getNextEntry();
-            while (entry != null) {
-                final Path target = root.resolve(entry.getName()).normalize();
-                if (!target.startsWith(root)) {
-                    throw new IOException(
-                        String.format(
-                            "Zip entry '%s' would write outside '%s'",
-                            entry.getName(), root
-                        )
-                    );
-                }
-                if (entry.isDirectory()) {
-                    Files.createDirectories(target);
-                } else {
-                    Files.createDirectories(target.getParent());
-                    Files.copy(zis, target, StandardCopyOption.REPLACE_EXISTING);
-                }
-                zis.closeEntry();
-                entry = zis.getNextEntry();
-            }
-        }
     }
 
     /**
